requirements.txt
# CORE BACKEND & ASYNC
fastapi~=0.118.0
uvicorn[standard]~=0.37.0
python-dotenv~=1.1.1
pydantic~=2.11.9

# LLM & ORCHESTRATION (Google Gemini)

# RAG & LANGCHAIN

# POSTGRESQL & DATABASE (For Metadata and Logs)
sqlalchemy~=2.0.43

# AGENT TOOLS


#Upload, Play, Loader, FileText, CheckCircle, XCircle from lucide-react


psycopg2~=2.9.11
pandas~=2.3.2